+++ src/front/pages/Home.jsx
@@ -1,92 +1,15 @@
 import React, { useEffect } from "react"
-import rigoImageUrl from "../assets/img/rigo-baby.jpg";
 import useGlobalReducer from "../hooks/useGlobalReducer.jsx";
 import MapVisual from "../components/MapVisual.jsx";
 
 
 export const Home = () => {
 
-	const { store, dispatch } = useGlobalReducer()
-	const foodTrucks = [
-		{
-			name: "The Rolling Taco",
-			address: "123 Main St, Austin, TX 78701"
-		},
-		{
-			name: "Burger Bus",
-			address: "456 Elm St, Denver, CO 80220"
-		},
-		{
-			name: "Curry Cruiser",
-			address: "789 Oak Ave, San Jose, CA 95110"
-		}
-	];
-
-	const loadMessage = async () => {
-		try {
-			const backendUrl = import.meta.env.VITE_BACKEND_URL
-
-			if (!backendUrl) throw new Error("VITE_BACKEND_URL is not defined in .env file")
-
-			const response = await fetch(backendUrl + "/api/hello")
-			const data = await response.json()
-
-			if (response.ok) dispatch({ type: "set_hello", payload: data.message })
-
-			return data
-
-		} catch (error) {
-			if (error.message) throw new Error(
-				`Could not fetch the message from the backend.
-				Please check if the backend is running and the backend port is public.`
-			);
-		}
-
-	}
-
-	useEffect(() => {
-		loadMessage()
-	}, [])
+	
 
 	return (
-		<div className="text-center mt-5">
-			<h1 className="display-4">Hello Rigo!!</h1>
-			<p className="lead">
-				<img src={rigoImageUrl} className="img-fluid rounded-circle mb-3" alt="Rigo Baby" />
-			</p>
-			{/* <MapVisual/> */}
-			<MapVisual
-				FoodTruckName="The Baconator"
-				address="1600 Amphitheatre Parkway, Mountain View, CA"
-			/>
-			<MapVisual
-				FoodTruckName="The Silent Killer"
-				address="10600 Okeechobee Blvd, West Palm Beach, FL"
-			/>
-			<MapVisual
-				FoodTruckName="The Roadhouse"
-				address="21699-A FL-7, Boca Raton, FL 33428"
-			/>
-
-			{foodTrucks?.map((truck, index) => (
-				<div key={index}>
-					<MapVisual
-						FoodTruckName={truck.name}
-						address={truck.address}
-					/>
-				</div>
-
-			))}
-
-			<div className="alert alert-info">
-				{store.message ? (
-					<span>{store.message}</span>
-				) : (
-					<span className="text-danger">
-						Loading message from the backend (make sure your python 🐍 backend is running)...
-					</span>
-				)}
-			</div>
+		<div>
+			<h1>Home</h1>
 		</div>
 	);
 };
